+++ tests/Uri/Filtering/ExpressionFilterParserTest.php
@@ -38,13 +38,14 @@ class ExpressionFilterParserTest extends TestCase
 
     public function testParse()
     {
-        $_SERVER["REQUEST_URI"] = "/getter?filter=stringProperty eq 'string' and contains(stringProperty,'asdf') and intProperty in (1,2,3) or boolProperty ne true and relation.property eq null";
+        $_SERVER["REQUEST_URI"] =
+            "/getter?filter=stringProperty eq 'string' and contains(stringProperty,'asdf') and intProperty in (1,2,3) or boolProperty ne true and relation.property eq null and stringProperty eq datetime'2018-12-01'";
         $request = ServerRequestFactory::createFromGlobals();
         $up = new UriParser($request, self::$mr);
         $parser = new ExpressionFilterParser(new DoctrineQueryExpressionBuilder(self::$mr, $up->getPath()));
         $up->setFilterParser($parser);
         $this->assertEquals(
-            "((getter.stringProperty = 'string' AND getter.stringProperty LIKE '%asdf%') AND getter.intProperty IN(1, 2, 3)) OR (getter.boolProperty <> true AND relation.property IS NULL)",
+            "((getter.stringProperty = 'string' AND getter.stringProperty LIKE '%asdf%') AND getter.intProperty IN(1, 2, 3)) OR ((getter.boolProperty <> true AND relation.property IS NULL) AND getter.stringProperty = '2018-12-01T00:00:00+00:00')",
             (string)$up->getFilter()->getCondition()
         );
         $this->assertArrayHasKey('relation', $up->getFilter()->getRequiredJoins());
@@ -56,13 +57,13 @@ class ExpressionFilterParserTest extends TestCase
 
     public function testDoctrineCriteriaExpression()
     {
-        $url = "stringProperty eq '3' and intProperty in (1,2,3) or boolProperty ne true";
+        $url = "stringProperty eq '3' and intProperty in (1,2,3) or boolProperty ne true and stringProperty eq datetime'2018-12-01'";
         $parser = new ExpressionFilterParser(new DoctrineCriteriaExpressionBuilder());
         $parser->parse($url);
         $visitor = new QueryExpressionVisitor(['t']);
         $result = $visitor->dispatch($parser->getCondition());
         $this->assertEquals(
-            "(t.stringProperty = :stringProperty AND t.intProperty IN(:intProperty)) OR t.boolProperty <> :boolProperty",
+            "(t.stringProperty = :stringProperty AND t.intProperty IN(:intProperty)) OR (t.boolProperty <> :boolProperty AND t.stringProperty = :stringProperty_3)",
             (string)$result
         );
     }